Digit replacement corrected
The replacement of digits in patterns resulted in some wrong results.
Now, multiple digits are no more replaced in group, but every single
digit. This results in more pattern searches in the first search step,
but less searches in the second step, and results in about the same
search time.