Adding full sub_key generation for iniParts
I noticed that in the Small file tests of browscap, it seemed like some
tests would stutter a bit.  Turned out that because the small file has
less data, not as many files are created.  When fetching the iniPart
for a matched pattern, if the cache file doesn’t exist, the inipart
cache is rebuilt.  The createPatterns method has a safe-guard for this
by creating all possible cache files, even if they don’t contain data.
This adds a similar method for the inipart cache.

When running the tests with this in place, the visible stutter is gone,
and the test suite runs a bit faster (~2.3 minutes vs. ~2.7 minutes).
The savings would be more substantial if the large or standard runs
attempted to lookup a hash that didn’t have a matching file, but I
think they make enough files that this doesn’t happen (the Lite file
was the only one that caused the “file does not exist” error before the
file_exists check was added).

This will cause a lot of empty cache files to be created, since the
inipart files have been bumped to 3 characters, but most of them will
be empty.